@@ -6,31 +6,31 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
-  mobx: ^2.0.1
-  flutter_mobx: ^2.0.0
-  flutter_modular: ^3.1.0  
-  google_fonts: ^2.0.0
-  oktoast: ^3.0.0
-  firebase_auth: ^1.1.0
-  cloud_firestore: ^1.0.5
-  firebase_core: ^1.0.3
-  google_sign_in: ^5.0.1
-  simple_animations: ^3.0.3
-  flutter_svg: ^0.21.0+1
+  cupertino_icons: ^1.0.5
+  mobx: ^2.0.7+3
+  flutter_mobx: ^2.0.6+1
+  flutter_modular: ^5.0.3  
+  google_fonts: ^3.0.1
+  oktoast: ^3.2.0
+  firebase_auth: ^3.3.20
+  cloud_firestore: ^3.1.18
+  firebase_core: ^1.18.0
+  google_sign_in: ^5.3.3
+  simple_animations: ^4.1.0
+  flutter_svg: ^1.1.0
   flutter_feather_icons: ^2.0.0+1
   flutter_staggered_animations: ^1.0.0
   
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mobx_codegen: ^2.0.1+3 
-  build_runner: ^1.12.2     
+  mobx_codegen: ^2.0.7 
+  build_runner: ^2.1.11     
 
 flutter:
 
