@@ -3,11 +3,11 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:pautas/src/app/theme/text_styles.dart';
 
 class DetailStep extends StatelessWidget {
-  DetailStep({Key key, this.title = "", this.text, this.image}) : super(key: key);
+  DetailStep({Key? key, this.title = "", this.text, this.image}) : super(key: key);
 
   final String title;
-  final String text;
-  final String image;
+  final String? text;
+  final String? image;
 
   final TextStyles stylesConsts = TextStyles();
 
@@ -24,7 +24,7 @@ class DetailStep extends StatelessWidget {
         Expanded(
           child: Container(
             margin: const EdgeInsets.only(left: 20),
-            child: SvgPicture.asset(image,
+            child: SvgPicture.asset(image!,
               width: MediaQuery.of(context).size.height * 0.5,
               height: MediaQuery.of(context).size.width * 0.9,
             )
@@ -32,7 +32,7 @@ class DetailStep extends StatelessWidget {
         ),
         Container(
           margin: EdgeInsets.only(left: 5, right: 5, top: 5),
-          child: Text(text,
+          child: Text(text!,
             style: stylesConsts.loginHint,
             textAlign: TextAlign.center,
             overflow: TextOverflow.ellipsis,
